Remove now unused EVM extension
