+++ runtime/Cargo.toml
@@ -55,7 +55,7 @@ pallet-democracy = { git = "https://github.com/paritytech/substrate", default-fe
 pallet-scheduler = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "rococo-v1" }
 pallet-collective = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "rococo-v1" }
 
-moonbeam-extensions-evm = { path = "extensions/evm", default-features = false }
+moonbeam-tracing-hook = { path = "tracing_hook", default-features = false }
 moonbeam-rpc-primitives-debug = { path = "../primitives/rpc/debug", default-features = false }
 moonbeam-rpc-primitives-txpool = { path = "../primitives/rpc/txpool", default-features = false }
 
@@ -114,7 +114,7 @@ std = [
 	"pallet-scheduler/std",
 	"pallet-collective/std",
 	"author-inherent/std",
-	"moonbeam-extensions-evm/std",
+	"moonbeam-tracing-hook/std",
 	"parachain-info/std",
 	"cumulus-pallet-parachain-system/std",
 	"cumulus-primitives-core/std",
